Colocate all chat log event handlers
